src/requirements.txt
aiohttp
motor
aiohttp_cors
pyTelegramBotAPI
python-dotenv
pyyaml
sentry-sdk
aiohttp-swagger[perfomance]
pillow
jsonpickle
openai
langchain
langchain-community
langchain-openai
pydub
tiktoken
lxml
pandas
faiss-cpu
pypdf
bs4
html2text